+++ src/compiler/seman/type/SemType.java
@@ -24,6 +24,6 @@ public abstract class SemType {
 	 * 
 	 * @return Velikost podatkovnega tipa v bytih.
 	 */
-	public abstract int size();
+	public abstract long size();
 
 }
